Removed old parser and `convert-from-json`
