+++ kmir/src/kmir/__main__.py
@@ -14,7 +14,7 @@ from pyk.proof.reachability import APRProof, APRProver
 from pyk.proof.tui import APRProofViewer
 
 from kmir.build import HASKELL_DEF_DIR, LLVM_LIB_DIR, haskell_semantics, llvm_semantics
-from kmir.convert_from_definition.v2parser import parse_json
+from kmir.convert_from_definition.parser import parse_json
 from kmir.kmir import KMIR, KMIRAPRNodePrinter
 
 if TYPE_CHECKING:
